+++ app/src/main/java/com/moneyguardian/ui/ItemPagosFragment.java
@@ -25,13 +25,13 @@ public class ItemPagosFragment extends Fragment {
     private static final String USERS_AND_PAYMENTS = "balance";
 
     private String nombreItem;
-    private HashMap<UsuarioParaParcelable,Integer> mapBalance = new HashMap<>();
+    private HashMap<UsuarioParaParcelable,Double> mapBalance = new HashMap<>();
 
     private TextView tvNombreItem;
     private RecyclerView rvBalance;
 
     public static ItemPagosFragment newInstance(String param1,
-                                                HashMap<UsuarioParaParcelable,Integer> param3) {
+                                                HashMap<UsuarioParaParcelable,Double> param3) {
         ItemPagosFragment fragment = new ItemPagosFragment();
         Bundle args = new Bundle();
         args.putString(NAME, param1);
@@ -46,7 +46,7 @@ public class ItemPagosFragment extends Fragment {
         super.onCreate(savedInstanceState);
         if (getArguments() != null) {
             nombreItem = getArguments().getString(NAME);
-            mapBalance = (HashMap<UsuarioParaParcelable, Integer>) getArguments().get(USERS_AND_PAYMENTS);
+            mapBalance = (HashMap<UsuarioParaParcelable, Double>) getArguments().get(USERS_AND_PAYMENTS);
         }
     }
